[#15] feat: Add support for .NET Core and .NET Standard
- conditional compilation is used to keep .NET Framework specific behavior that fixed a bug related to Windows Authentication (see mentioned PiWeb issue in RestClientBase)
