+++ src/Api.csproj
@@ -1,12 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   
   <PropertyGroup>
-    <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
     <EnableDefaultEmbeddedResourceItems>false</EnableDefaultEmbeddedResourceItems>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <ProjectTypeGuids>{60dc8134-eba5-43b8-bcc9-bb4bc16c2548};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
+    <ProjectTypeGuids>{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <RootNamespace>Zeiss.IMT.PiWeb.Api</RootNamespace>
-    <TargetFramework>net472</TargetFramework>
+    <TargetFrameworks>netstandard2.0;netcoreapp3.0;net48</TargetFrameworks>
   </PropertyGroup>
 
   <PropertyGroup Label="Assembly information">
@@ -24,7 +23,7 @@
     <Product>ZEISS PiWeb API</Product>
   </PropertyGroup>
 
-  <PropertyGroup Label="Nuget package specifications">
+  <PropertyGroup Label="NuGet package specifications">
     <AllowedOutputExtensionsInPackageBuildOutputFolder>$(AllowedOutputExtensionsInPackageBuildOutputFolder);.pdb; .xml;</AllowedOutputExtensionsInPackageBuildOutputFolder>
     <Authors>$(Company)</Authors>
     <EmbedAllSources>true</EmbedAllSources>
@@ -41,39 +40,28 @@
   <PropertyGroup>
     <DebugType>Full</DebugType>
   </PropertyGroup>
-  
-  <ItemGroup Label=".NET references">
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.IdentityModel" />
-    <Reference Include="System.Net" />
-    <Reference Include="System.Net.Http" />
+
+  <ItemGroup Label=".NET references" Condition="'$(TargetFrameworkIdentifier)' == '.NETFramework'">
     <Reference Include="System.Net.Http.WebRequest" />
-    <Reference Include="System.Runtime.Serialization" />
-    <Reference Include="System.Security" />
-    <Reference Include="System.ServiceModel" />
-    <Reference Include="System.Xml" />
   </ItemGroup>
-  
-  <ItemGroup Label="Nuget references">
-    <PackageReference Include="IdentityModel" Version="1.13.0" />
+    
+  <ItemGroup Label="NuGet references">
+    <PackageReference Include="IdentityModel" Version="2.0.0" />
     <PackageReference Include="JetBrains.Annotations" Version="2018.2.1" />
     <PackageReference Include="Microsoft.IdentityModel.Logging" Version="5.2.1" />
     <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="5.2.1" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
     <PackageReference Include="Newtonsoft.Json.Bson" Version="1.0.2" />
     <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="5.2.1" />
+    <PackageReference Include="System.Resources.Extensions" Version="4.7.0" />
   </ItemGroup>
   
   <ItemGroup>
-  <None Include="license.txt" Pack="true" PackagePath="$(PackageLicenseFile)" />
+    <None Include="license.txt" Pack="true" PackagePath="$(PackageLicenseFile)" />
   </ItemGroup>
 
   <ItemGroup>
-    <EmbeddedResource Include="**/*.resx" />
-    <EmbeddedResource Update="**/*.resx">
+    <EmbeddedResource Include="**/*.resx">
       <DependentUpon Condition="Exists('$([System.String]::Copy(&quot;%(RelativeDir)%(FileName)&quot;).Replace(&quot;.de&quot;, &quot;&quot;)).cs')">$([System.String]::Copy('%(FileName)').Replace('.de', '')).cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
